requirements.txt
google-generativeai
pillow
pdf2image
python-dotenv
supermemory

